Worked on device number support #9

+++ libfshfs/libfshfs_file_entry.h
@@ -243,8 +243,8 @@ int libfshfs_file_entry_get_group_identifier(
 LIBFSHFS_EXTERN \
 int libfshfs_file_entry_get_device_number(
      libfshfs_file_entry_t *file_entry,
-     uint8_t *major_device_number,
-     uint8_t *minor_device_number,
+     uint32_t *major_device_number,
+     uint32_t *minor_device_number,
      libcerror_error_t **error );
 
 LIBFSHFS_EXTERN \
